changing dockerfiles to no longer need renaming
